Refactor data access to use Dapper and new interfaces
This commit refactors the data access layer by replacing the `Repository<Credentials>` with an `ICredentialsRepository` interface, allowing for a more flexible implementation using Dapper instead of Entity Framework.

Key changes include:
- Updated `AccountOps.cs` and `PasswordOps.cs` to utilize the new repository pattern.
- Changed serialization layouts in `CredentialsPacket.cs`, `CredsUpdatePacket.cs`, and `ResponsePacket.cs` to `Explicit`.
- Modified `CommandExtensions.cs` to return `UInt32` for packet magic numbers.
- Enhanced logging in `Registry.cs` for packet catalog registration.
- Updated `LoginSence.cs` to handle user credentials with encrypted packets.
- Introduced a new `ParallaxLayer` class in `MainScene.cs` for improved visuals.
- Refactored `Program.cs` for manual service composition.
- Added `DbConnectionFactory`, `DbConfig`, and `DbInitializer` for database management.
- Removed `GameDbContext.cs`, `GameDbContextFactory.cs`, and `Repository.cs` to streamline the architecture.
- Updated `Nalix.Infrastructure.csproj` to include Dapper and SQLite package references.

+++ src/Nalix.Communication/Extensions/CommandExtensions.cs
@@ -16,5 +16,5 @@ public static class CommandExtensions
     /// </summary>
     /// <param name="command">Giá trị enum OpCommand.</param>
     /// <returns>Giá trị ushort tương ứng.</returns>
-    public static System.UInt16 AsUInt32(this PacketMagic command) => (System.UInt16)command;
+    public static System.UInt32 AsUInt32(this PacketMagic command) => (System.UInt32)command;
 }
